Update change log and refactor.
